+++ changelog.md
@@ -13,6 +13,7 @@
 - Replace unmaintained `mergedeep` dependency by `deepmerge`.
 - Expose `click._utils.UNSET` sentinel as `click_extra.UNSET`.
 - Fix detection in the `--version` option of the module in which the user's CLI is implemented.
+- Rename `click_extra.tabulate` namespace to `click_extra.table`.
 - Mark Python 3.14-dev tests as unstable again, because of `PyYAML`.
 
 ## [5.1.1 (2025-08-24)](https://github.com/kdeldycke/click-extra/compare/v5.1.0...v5.1.1)